pyproject.toml
[tool.poetry]
name = "python-dapr-demo"
version = "0.1.0"
description = ""
authors = ["Aaron Blondeau <aaronblondeau@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
dapr = "^1.13.0"
dapr-ext-fastapi = "^1.12.0"
uvicorn = "^0.30.1"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
